Join tests logic
